chore: renaming and small refactoring of promotionboxbadge
