+++ lib/src/promotion_box/promotion_box.assets.dart
@@ -1,4 +1,4 @@
-part of 'promotion_box.dart';
+part of 'sbb_promotion_box.dart';
 
 sealed class _PromotionBoxAssets {
   static const noise = 'packages/sbb_design_system_mobile/lib/assets/noise.png';
